Change password strength meter to stimulus
